backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai-whisper==20231117
python-multipart==0.0.9
ffmpeg-python==0.2.0